+++ swarmlib/cuckoosearch/nest.py
@@ -3,7 +3,10 @@
 #  Licensed under the BSD 3-Clause License. See LICENSE.txt in the project root for license information.
 # ------------------------------------------------------------------------------------------------------
 
+from copy import deepcopy
 from typing import Tuple
+
+import numpy as np
 
 from ..util.coordinate import Coordinate
 
@@ -12,17 +15,22 @@ class Nest(Coordinate):
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        self.__generated = True
+        self.__abandoned = True
 
     @property
-    def generated(self) -> bool:
+    def abandoned(self) -> bool:
         """
-        Indicates whether this nest was created randomly before (True) or updated (False)
+        Indicates whether a nest was abandoned or not
 
         Returns:
-            bool: True or False
+            bool: True the nest was abandoned otherwise False
         """
-        return self.__generated
+        return self.__abandoned
+
+    def abandon(self) -> None:
+        self.__abandoned = True
+        self._position = np.random.uniform(self._lower_boundary, self._upper_boundary, 2)
+        self._value = self._function(self._position)
 
     def update_pos(self, new_position: Tuple[float, float]) -> None:
         """
@@ -32,9 +40,9 @@ class Nest(Coordinate):
             new_position {Tuple[float, float]} -- The new position
         """
 
-        if self.__generated:
-            self.__generated = False
         new_value = self._function(new_position)
         if new_value < self._value:
+            if self.__abandoned:
+                self.__abandoned = False
             self._value = new_value
-            self._position = new_position
+            self._position = deepcopy(new_position)