+++ swarmlib/cuckoosearch/visualizer.py
@@ -16,14 +16,13 @@ class Visualizer(BaseVisualizer):
         ax = self._fig.gca(label='BaseAxis')
         self.__best_nests_artist, = ax.plot([], [], 'o', color='#ffff00', ms=6)
         self.__best_nests = [[], []]
-        self.__calculate_velocity = []
+        self._abandon_map = []
 
     def add_data(self, **kwargs) -> None:
         super().add_data(**kwargs)
-        # Indicates for which nest shall be a velocity calculated (True) or not (False)
-        # Used to avoid velocity calculation for new generated nests.
-        calculate_velocity = ~np.array(kwargs['generated'])
-        self.__calculate_velocity.append(np.array([calculate_velocity, calculate_velocity]))
+        # Indicates whether the nest was generated this iteration or not
+        abandoned = np.array(kwargs['abandoned'])
+        self._abandon_map.append(np.array([abandoned, abandoned]))
 
         x_pos, y_pos = kwargs['best_position']
         self.__best_nests[0].append(x_pos)
@@ -31,7 +30,7 @@ class Visualizer(BaseVisualizer):
 
     def replay(self):
         # Prepare velocities before starting replay
-        self._velocities = [(self._positions[index+1]-position)*self.__calculate_velocity[index+1] for index, position in enumerate(self._positions[:-1])]
+        self._velocities = [self._positions[index+1]-position for index, position in enumerate(self._positions[:-1])]
         self._velocities.insert(0, np.zeros(self._positions[0].shape))
         self._velocities.append(np.zeros(self._positions[0].shape))
 
@@ -44,9 +43,12 @@ class Visualizer(BaseVisualizer):
         return [*base_artists, self.__best_nests_artist]
 
     def _animate(self, i: int, frames: int):
+        if self._index < len(self._abandon_map)-1:
+            # Color the velocity different when the nest is abandoned
+            self._vel_color = np.where(self._abandon_map[self._index+1].sum(axis=0), '#373737', '#CFCFCF')
         base_artists = super()._animate(i, frames)
 
-        self.__best_nests_artist.set_data(self.__best_nests[0][:self._index], self.__best_nests[1][:self._index])
+        self.__best_nests_artist.set_data(self.__best_nests[0][:self._index+1], self.__best_nests[1][:self._index+1])
         self.__best_nests_artist.set_markersize(self._marker_size)
 
         return [*base_artists, self.__best_nests_artist]